Terminado en totalidad Arribo
