+++ datatypes/headers/DtArribo.h
@@ -1,29 +1,29 @@
 #ifndef DTARRIBO_H
 #define DTARRIBO_H
 /* Esto es nada mas para estructura */
-#include "../../datatypes/headers/DtFecha.h"
-#include "../../datatypes/headers/DtBarco.h"
+#include "DtFecha.h"
+#include "DtBarco.h"
 
 class DtArribo
 {
 private:
     DtBarco barcoQueArriba;
     DtFecha fechaDeArribo;
-    float cantidadDespachada;
+    float cargaDespachada;
 
 public:
     // Constructor
-    DtArribo(DtBarco barcoQueArriba, DtFecha fechaDeArribo, float cantidadDespachada);
+    DtArribo(DtBarco barcoQueArriba, DtFecha fechaDeArribo, float cargaDespachada);
 
     // Getters
     DtBarco getBarcoQueArriba();
     DtFecha getFechaDeArribo();
-    float getCantidadDespachada();
+    float getCargaDespachada();
 
     // Setters
-    void setBarcoQueArriba(DtBarco barcoQueArriba);
-    void setFechaDeArribo(DtFecha getFechaDeArribo);
-    void setCantidadDespachada(float cantidadDespachada);
+    void setBarcoQueArriba(DtBarco paramBarcoQueArriba);
+    void setFechaDeArribo(DtFecha paramFechaDeArribo);
+    void setCargaDespachada(float paramCargaDespachada);
 
     virtual ~DtArribo();
 };
